requirements.txt
python-chess>=1.9.0 